requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
alembic==1.12.1
jinja2==3.1.2
python-dotenv==1.0.0
reportlab==4.0.7

# System Requirements
# ------------------
# PostgreSQL 14+ with pg_cron extension
# Installation on macOS:
#   brew install postgresql@14
#   brew services start postgresql@14
#
# Add to ~/.zshrc:
#   export PATH="/opt/homebrew/opt/postgresql@14/bin:$PATH"
#
# Or install Postgres.app from https://postgresapp.com/
